added some extra buttons. cleaned up...
added some extra buttons for copy paste and profile. 
Cleaned up some of the bad code i wrote last time. its about as fast as before but in extreme cases it can still be slow. 
Have some ideas as to how to speed it up but most of it is because of the chroma library used.